refactor: move window management
